requirements.txt
dhanhq>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
ta>=0.11.0
pytest>=7.4.0
